pumpkin: workaround SoC bootloop
If we let modprobe -a loads the wifi modules at boot time, there is
high chance of the SoC to reset, maybe due to a brownout reset. This
commit blacklist the wifi modules in order to delay the loading of the
module later in boot, and avoid this reset.
This workaround is applied *only* for when sysvinit is used because
I could not reproduce the issue with systemd since the modprobe since
to be done a little bit later.

Signed-off-by: Fabien Parent <fparent@baylibre.com>