u-boot-fw-utils: drop bbappend
The bbappend file for the fw-utils recipe gets in the way of another
bbappend we want to have in meta-baylibre-mender to override the
fw_env.config file. Make it into an inc.

Also: convert the SRC_URI += to SRC_URI_append or otherwise the file
in the upper layer won't be picked up.

Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>