tf-a-priv: make BL32/OP-TEE optional
Add a new DISTRO_FEATURES option in order to add optee to the image.
Depending on whether the optee feature is available, we will build BL2
with support or not of BL32.

We still need optee for the non private version of tf-a because we
are just shipping a binary that require optee.

Signed-off-by: Fabien Parent <fparent@baylibre.com>