kernel: make recipe include linux-yocto
Include linux-yocto in the kernel recipes in order to take advantage
of the feature provided by linux-yocto.

Remove the LICENSE field since it is also provided by linux-yocto.
And change the LINUX_VERSION_EXTENSION to not be an append anymore
in order to override the extension provided by linux-yocto and use
our own simple extension: -mtk.

Signed-off-by: Fabien Parent <fparent@baylibre.com>