kernel: use new mtk-v4.19 branch
Switch to the new mtk-v4.19 branch in order to support other SoC.
Update the kernel to v4.19.60 at the same time.

Starting with the mtk-v4.19 branch, we are using the vesper dtbo since
there is no vesper device tree anymore in the kernel source code. To
not break anything we add some environment configuration to u-boot
in order to load by default the vesper cape if vesper-hat is in the
MACHINE_FEATURES.

Signed-off-by: Fabien Parent <fparent@baylibre.com>