tf-a: update hash
Update to the latest code for TF-A, this includes the need to define
a new TFA_BOARD_NAME variable that is used to select the correct
EMI setting for the board on pumpkin.

In addition the RTC code for MT8516/MT8167 has been added to bl2 in
order to have the RTC working on pumpkin.

Signed-off-by: Fabien Parent <fparent@baylibre.com>