u-boot: simplify loading of overlays
This commit introduces a new variable in order to simplify the way
kernel device tree overlays are automatically loaded at boot time.

Until now only rpi-display.dtbo and vesper.dtbo can be automatically
loaded at boot time. We want to be able to automatically load more
device tree overlays without having to add more and more code
in the u-boot recipe to handle all the dtbo. This commits now
define a new variable called KERNEL_DEVICETREE_OVERLAYS_AUTOLOAD
that specifies which overlay should be automatically loaded.

This new variable should be used in local.conf but for rpi-display.dtbo
and vesper.dtbo we keep compatibility with the current build by
automatically adding the overlays to the AUTOLOAD variable if the
corresponding DISTRO_FEATURES is set.

Signed-off-by: Fabien Parent <fparent@baylibre.com>