tf-a-priv: update to pick up secure code
The private TF-A repository is now building the secure code. Update
the hash to pick up that change. This means that technically we can
share binaries that were built from the private TF-A repository.

Note: the TF-A private source should not be shared with anyone outside
of BayLibre or MediaTek.

Signed-off-by: Fabien Parent <fparent@baylibre.com>