conf: machine: add PREFERRED_VERSION for weston
Since we added weston 4.0 in the layer, it will also be the one used
by default since meta-mediatek-bsp layer has a higher priority over
poky. For mt8167 we still want the latest supported weston release.

This commit will force the usage of a specific version of weston
for mt8167 & mt8183.

Signed-off-by: Fabien Parent <fparent@baylibre.com>