conf: move DISTRO_FEATURES_remove into machine files
Not all DISTRO_FEATURES are supported by all the machines, for
instance mt8183 does not yet support x11 nor vulkan. It was defined
in the local.conf to remove these DISTRO_FEATURES from poky's default,
but when these change, people have to modify their local.conf which
is not convenient.

Remove the unsupported DISTRO_FEATURES in the machine configs instead.

Signed-off-by: Fabien Parent <fparent@baylibre.com>