dtbo: add an audio overlay for i300-pumpkin
Add a new audio overlay using i2s in and out on the 40pins header.
Also add its corresponding alsa and pulse configurations.
In the future, it will also be possible to add tdm in and spdif
in/out to this pin header configuration.
This new config is selected when the variable I300_PUMPKIN_AUDIO_CONF
contains the keyword "i2s".
In the overlay, use the flag GPIO_ACTIVE_LOW for the gpio controlling
the I2S/SDCARD mux. In that case the gpio will be high when active-low
is set, which is what we want to activate the I2S.

Signed-off-by: Nicolas <nbelin@baylibre.com>
Signed-off-by: Nicolas Belin <nbelin@baylibre.com>