alsa: add a jack microphone pcm in asound.conf when using internalmic
Add a jack microphone pcm to the asound.conf file in case we want to
use the jack microphone instead of the pdm mics.
The difference with the pdm mics is that the jack mic is only 1 channel.
Please Note that to use the jack mic you will have to change a couple of
alsa settings:
amixer set -c mtsndcard 'AIF TX Mux',0 'Analog MIC'
amixer set -c mtsndcard 'Left PGA Mux',0 'CH1'

Signed-off-by: Nicolas <nbelin@baylibre.com>