ci: fail CI on warnings
The is two rules related to the documentation:
* build_pages: which just only checks that the documentation builds fine
* pages: build and deploy the pages for the stable branch.

For the `build_pages` rule, we want to fail on any errors. To make that
happens we add two options to the sphinx-build command line:
	-n: nit-picky mode, warn about all missing references
	-W: turn warnings into errors

For the pages target, we want it to always succeed when possible so
we don't enable these two options.

Signed-off-by: Fabien Parent <fparent@baylibre.com>