alsa: add the alsa configs for i350-sb35 and i350-evk
Add the asound.conf and the asound.state files for
both i350-sb35 and i350-evk.
By default jack in/out is used. Dmic needs to be
mentioned explicitaly to be used.
On i350-evk, Line Out is used at the same time as
jack out by default.
On i350-evk, Audio_MicSource1_Setting needs to be set
to ADC1 in order to use the AMIC instead of jack in.

Signed-off-by: Nicolas Belin <nbelin@baylibre.com>