optee: fix build by adding support for optee-os-tadevkit
The build is broken since meta-arm added the following commit:
d69ae1e0008f optee: update optee-os.inc to support external TAs

The issue is that OP-TEE now depends on this packages and this packages
requires some variable to be set to be able to be built properly.

optee-os-tadevkit is rebuilding op-tee OS to it needs the same variable
definitions, so let move everything from optee-os recipe to a inc
file that will be included for both optee-os and optee-os-tadevkit.

Signed-off-by: Fabien Parent <fparent@baylibre.com>