optee-os: rework early TA integration
The way optee-otp is integrated works but will not scale well if other
TAs end up being necessary at the early stage.

With this rework, the early TA list should be provided through
MACHINE_OPTEE_EARLY_TA variable. By adding this to DEPENDS of optee-os,
we make sure that early TAs ends up in the sysroot of optee-os which is
then used to collect the early TAs.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>