aiotflash: modernize the way to make a package for aiot-flash
The initial method to pack the image for aiot-flash did not handle the
dependencies so well. This leads (weak) distro to do all kind of funny
things.

Instead of relying on an extra IMAGE class, this add a new image type,
directly dependent on wic.img.

To enable this, just IMAGE_FSTYPES += "aiotflash.tar"

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>