Upgrade clinfo to 3.0.21.02.21 via bbappend
Clinfo recipe on openembedded.org [1] has not been updated since 2018.
As a result of this, clinfo does not show correct information pertaining to OpenCL 3.0 released in Sept 2020 [2].

As a solution to this, a bbappend file is created that changes SRCREV to 3.0.21.02.21 tag [3]

After this upgrade, the output of `clinfo` shows that OpenCL version is 3.0 rather than 2.0 in the previous version.

[1]: http://cgit.openembedded.org/meta-openembedded/tree/meta-oe/recipes-support/opencl/clinfo_2.2.18.04.06.bb?h=master
[2]: https://github.com/KhronosGroup/OpenCL-Docs/releases/tag/v3.0.4
[3]: https://github.com/Oblomov/clinfo/commit/d34bc1a3bdc148e2e1fe64998398e1a0552ab04c

Signed-off-by: Suhrid Subramaniam <suhrid.subramaniam@mediatek.com>