optee-ewriter: MediatekPropritary is unknown
MediatekPropritary does not match any known license.
This triggers warning while building any image featuring this package.

It is not possible for the user to make an informed choice without the
license. It is also not possible to ensure it is actually compatible with
BSD 2 license terms. Until this MediatekPropritary gets documented, move to
CLOSED license.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>