optee-ewriter: properly handle dependencies
Apparently everyone should want to ship optee-ewriter by default.
This is controversial choice. With the current configuration, even if
optee-ewriter is removed from the image, efuse support is still built in
optee-os. This makes little sense.

Clean-up the dependencies and requirements of optee-ewriter so we can
properly de-activate it, if need be.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>