Add kernel-image to MACHINE_ESSENTIAL_EXTRA_RDEPENDS to make basic images boot
Without this, basic images (core-image-minimal, core-image-base, core-image-weston)
doesn't boot when meta-mediatek-bsp is used as standalone with it's dependencies.

MACHINE_ESSENTIAL_EXTRA_RDEPENDS is documented as:
"the list of packages are essential for the machine to boot"

Without this, the fitImage boot files isn't installed in the main
root filesystem /boot directory, and u-boot fails to boot.

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>