reicpes-bsp: mtk-secure-boot-tools: Add support of efuse.cfg
efuse.cfg could be used with aiot-tools to flash the efuse.
This adds a way to generate it. This could support 4 keys.
Instead of using EFUSE_KEY, we use EFUSE_KEY0 to EFUSE_KEY3.
For legacy, we can still provide EFUSE_KEY

This also update the hash to include the following commit:
d7e6ac6 Add support of efuse.cfg

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>