mtk-secure-boot-tools: Provide a way select the key to sign bl2 and lk
Currently, we use the key0 to sign lk.
This adds a new variable SBC_KEY that allows to use any of the four
keys supported. The variable EFUSE_KEY{0,1,2,3} are only used to
generate the config files.

This also replace the DA_KEY by the SBC_KEY.
The DA_KEY is never written in SoC and the ROM code actually only
use EFUSE_KEY{0,1,2,3} to check the signature.

Also update the hash to bring the following commit:
6643370 Fix a build issue
005cbd4 Add a way to select the key used to sign bootloaders

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>