gstreamer1.0-plugins-good: video: add colorimetry extensions
Gstreamer will generate a list of colorimetry, that contains
one default colorimetry for each colorspace, to speed up the
negotiation between source pad and sink pad.
But if the device requests a colorimetry which not on the list,
the pipeline will fail to establish the connection.

We made a way to append some colorimetry (for that we care)
via a predefined list.
It will only be appended if the device allows it.

Signed-off-by: Macross Chen <macross.chen@mediatek.com>