bsp: mtk-camisp-driver: Install ccd firmware
mtk-ccd is built within Remote Processor Framework. It requires a
firmware to start the driver. However, mtk-ccd bases on a software
instead of a firmware. That is why we need to add a dummy firmware for
the framework to load a firmware. We add it in the recipe because a
built-in firmware is ineffective in an OOT-driver.

Change-Id: I80651522ee1308a4ad4f3daff31c12d7935ab2e6
Signed-off-by: Chihhao Tsai <chihhao.tsai@mediatek.com>