dtbo: genio: Remove the 3.3v regulator always on of edp panel
Remove the edp 3.3v regulator always on for all platforms.

When suspending the platform with USB cable plugged in, regulator always
on would not power off the panel, then the HPD signal is not generated.
As a result, the panel could not be brought up after resume.

On the other hand, it seems that regulator-always-on avoids the panel
power sequence error setting. So adding the correct edp / lvds panel power
sequence delays [1] are needed to make the panels work normally.

[1]: I288a2dc28b9b3407925592232cc32ab9193d7896

Change-Id: I28c02f148a9535d2655292acd373ead332cfec95
CR-Id: AUTO00213194
Signed-off-by: xinlei lee <xinlei.lee@mediatek.com>
Signed-off-by: Tommy Chen <tommyyl.chen@mediatek.com>