recipes-kernel: linux: prepend kernel config for initramfs
The process bundle initramfs is based on kernel bbclass. For rootfs
security, this would ensure that initramfs related kernel configuration
is enabled correctly if the 'security' is in DISTRO_FEATURES.

This workaround is applied only for when 'security' in DISTRO_FEATURES
is used because there is the uncompress error for loading kernel from
fitImage. The possible error would be found as follows:
====
   Loading fdt from FIT Image at 4a000000 ...
...
   Verifying Hash Integrity ... sha256+ OK
   Booting using the fdt blob at 0x44000000
   Uncompressing Kernel Image
Error: inflate() returned -5
gzip compressed: uncompress error -1
====

Change-Id: I177058006b15043af5be3b1ba23cd38b99f2b597
Signed-off-by: Howard Lin <howard-yh.lin@mediatek.com>