multimedia: uvc-gadget: add uvc-gadget test program
This tool is from https://github.com/wlhe/uvc-gadget.

This tool is designed for testing the USB gadget function.
Currently, it is only used to output the image of a USB camera
to the UVC gadget 3.0 port. This allows the host PC to see the image
of the USB camera on the platform through UVC.

The following is a brief usage process:
1. Connect the USB camera and enable the UVC gadget.

2. Check the generated port:
`ls /sys/class/video4linux/ -l`, and you may see something like this:
..11290000.xhci/usb7/7-1/7-1.1/7-1.1:1.2/video4linux/video3
..11290000.xhci/usb7/7-1/7-1.1/7-1.1:1.2/video4linux/video4
..11201000.usb/gadget/video4linux/video5

3. Use the uvc-gadget tool to output the USB camera image to UVC:
`uvc-gadget -u /dev/video5 -v /dev/video3 -f 1 -r 1 -s 2`

4. After that, use a camera preview app on the host to see the UVC image.

Change-Id: Ieac5593137b3294c87dfa04d5b00bc7a1113fe88
Signed-off-by: Yow-Shin Liou <yow-shin.liou@mediatek.com>