linux-firmware: Add LICENCE.mediatek into firmware files
Add the LICENCE text for the SCP firmware binary files,
since these firmware binaries has a different license.

The license is the same as the one in:
https://git.kernel.org/pub/scm/linux/kernel/git/firmware/linux-firmware.git

Note that this file won't be used by the Yocto. Instead
Yocto refers to the license file located in the upstream
repo. We add the file here just for those who don't know
where to find the license.

Change-Id: I85ced07bb16fe59aa4342dccbc7fae9eb0e3ddb4
Signed-off-by: Pablo Sun <pablo.sun@mediatek.com>