v4l-utils: cec-ctl: resend standby in init_power_cycle_test
In the power cycle test, the first standby transmit must always succeed.
The issue arises from the code flow, which sends only one standby message
and then enters a loop to poll the power status. If this initial standby
message transmission fails, the first standby will therefore fail.

Modifying based on the latest cec-ctl.cpp on Github [1], the standby
message is resent if it fails. This involves sending one standby message
and one polling power status message in a loop, consequently improving
the success rate of the first standby.

Signed-off-by: tommyyl.chen <tommyyl.chen@mediatek.com>
CR-Id: AUTO00311036
[1]: https://github.com/gjasny/v4l-utils/blob/master/utils/cec-ctl/cec-ctl.cpp#L1386

Change-Id: I7fff2e85272ca3a894b1b4fd0ac5552e0e6b8dc0