u-boot-tools: Fix build failure in target mode
If CONFIG_VIDEO_LOGO or CONFIG_LCD_LOGO is enabled, headers will be
generated by running 'bmp_logo', which is built during the build
process. Although the generation is unnecessary (since it is only needed
when building u-boot), it has no harm when u-boot-tools is built in
native mode.

But this causes execution failure when building u-boot-tools in target
mode, because 'bmp_logo' now becomes a binary whose format is different
from the build host.

To fix the issue, we update the recipe to remove unneeded settings and
add a patch to remove the invocation.

We also update u-boot-common_git.inc to sync change from poky.

Change-Id: I5d8f2d0838eeaa7426fda8badbe7490232dddecc
Signed-off-by: Ramax Lo <ramax.lo@mediatek.com>